:recycle: refactor: 폴더구조 수정, 폰트 색상 적용
