+++ services/home/src/pages/JoinStudy/components/JoinStudyBox/style.ts
