+++ services/home/src/pages/JoinStudy/components/StudyInfo/index.tsx
